pr-sug: move to better namespace location
