+++ src/Arcus.Security.Core/SecretStoreBuilder.cs
