+++ src/Arcus.Security.Core/SecretStoreSource.cs
